+++ app/chat/components/friendBar.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Avatar, AvatarFallback} from "@/components/ui/avatar";
+import { Avatar, AvatarFallback } from "@/components/ui/avatar";
 import { useRouter } from "next/navigation";
 
 interface FriendBarProp {
