+++ app/chat/types/UserClass.ts
